% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowtype.R
\name{flowtype}
\alias{flowtype}
\title{flowtype.js for a better reading experience}
\usage{
flowtype(selector = "body", minimum = NULL, maximum = NULL,
  minFont = NULL, maxFont = NULL, fontRatio = NULL, width = 0,
  height = 0)
}
\description{
One and only function of this htmlwidget which runs
 \href{https://github.com/simplefocus/flowtype.js}{flowtype.js} to a
 \code{CSS selector} of your choice.  By default, \code{body} will be
 the \code{CSS selector}, so all text on the page will adjust to flowtype.
}
\examples{
\dontrun{
library("flowtypeR")
library("htmltools")

#  get NEWS from http://cran.r-project.org/doc/manuals/r-devel/NEWS.html
browsable(
  tagList(
    HTML(
      readLines("http://cran.r-project.org/doc/manuals/r-devel/NEWS.html")
    )
    ,flowtype()
  )
)

}
}

