% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowtype.R
\name{flowtype}
\alias{flowtype}
\title{flowtype.js for a better reading experience}
\usage{
flowtype(selector = "body", minimum = NULL, maximum = NULL,
  minFont = NULL, maxFont = NULL, fontRatio = NULL, width = 0,
  height = 0)
}
\arguments{
\item{selector}{\code{string} of a valid \code{CSS} selector.  \code{body} is the default.}

\item{minimum}{\code{integer} see \href{https://github.com/simplefocus/FlowType.JS#thresholds}{Thresholds}}

\item{maximum}{\code{integer} see \href{https://github.com/simplefocus/FlowType.JS#thresholds}{Thresholds}}

\item{minFont}{\code{integer} see \href{https://github.com/simplefocus/FlowType.JS#thresholds}{Thresholds}}

\item{maxFont}{\code{integer} see \href{https://github.com/simplefocus/FlowType.JS#thresholds}{Thresholds}}

\item{fontRatio}{\code{integer} see \href{https://github.com/simplefocus/FlowType.JS#font-size}{Font-size}}

\item{width}{a valid \code{CSS} size for the width of the htmlwidget container \code{div}.  The default
is \code{0} since we expect the container to be invisible.}

\item{height}{a valid \code{CSS} size for the height of the htmlwidget container \code{div}.  The default
         is \code{0} since we expect the container to be invisible.}
}
\description{
One and only function of this htmlwidget which runs
 \href{https://github.com/simplefocus/flowtype.js}{flowtype.js} on a
 \code{CSS selector} of your choice.  By default, \code{body} will be
 the \code{CSS selector}, so all text on the page will adjust to flowtype.
}
\examples{
\dontrun{
library("flowtypeR")
library("htmltools")

#  get NEWS from http://cran.r-project.org/doc/manuals/r-devel/NEWS.html
browsable(
  tagList(
    HTML(
      readLines("http://cran.r-project.org/doc/manuals/r-devel/NEWS.html")[8:145]
    )
    ,flowtype()
  )
)

}
}

